KokaCard and KokaCardEvent grows up!
Migrated the KokaCard and KokaCardEvent into classes.
The big change is that now they take a DocumentSnapshot and shows all relevant fields for that document.

 Changes to be committed:
	modified:   lib/pages/ContentViewerPage.dart
	modified:   lib/pages/InfoPage.dart
	modified:   lib/pages/NewsPage.dart
	modified:   lib/pages/ProgramPage.dart
	new file:   lib/widgets/organisms/KokaCard.dart
	new file:   lib/widgets/organisms/KokaCardEvent.dart
	deleted:    lib/widgets/organisms/kokaCard.dart
	deleted:    lib/widgets/organisms/kokaCardEvent.dart
