+++ lib/pages/NewsPage.dart
@@ -1,19 +1,15 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:oase/helpers/asset_helpers.dart';
-import 'package:oase/helpers/mapCategoryToColor.dart';
 import 'package:oase/pages/ContentViewerPage.dart';
 import 'package:oase/styles.dart';
-import 'package:oase/widgets/organisms/kokaCard.dart';
+import 'package:oase/widgets/organisms/KokaCard.dart';
 
 class NewsPage extends StatelessWidget {
   Widget _buildListItem(BuildContext context, DocumentSnapshot document) {
-    return kokaCard(
-      title: document['title'],
-      content: document['content'],
+    return KokaCard(
+      document: document,
       short: true,
-      colorStart: mapCategoryToStartColor(document['category'].toString()),
-      colorEnd: mapCategoryToEndColor(document['category'].toString()),
       onTapAction: () => Navigator.push(context,
           MaterialPageRoute(builder: (context) => ContentViewerPage(document))),
     );
